requirements.txt
appnope==0.1.4
asttokens==3.0.0
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.12
decorator==5.2.1
executing==2.2.0
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
git-filter-repo==2.47.0
ipykernel==6.29.5
ipython==9.0.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.1
packaging==24.2
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
pip==25.1
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.1
scipy==1.16.0
setuptools==78.1.1
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
wcwidth==0.2.13
wheel==0.45.1
